chore(app): remove atoms/buttons/OnDeviceDisplay folder and move buttons into atoms/buttons
remove atoms/buttons/OnDeviceDisplay folder and move buttons into atoms/buttons

RAUT-422
