+++ app/src/atoms/buttons/FloatingActionButton.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { FloatingActionButton } from '.'
+import { FloatingActionButton } from './'
 import type { Story, Meta } from '@storybook/react'
 import { ICON_DATA_BY_NAME } from '@opentrons/components/src/icons/icon-data'
