+++ app/src/atoms/buttons/FloatingActionButton.tsx
@@ -14,7 +14,7 @@ import {
   SPACING,
   TYPOGRAPHY,
 } from '@opentrons/components'
-import { StyledText } from '../../text'
+import { StyledText } from '../text'
 
 import type { IconName, StyleProps } from '@opentrons/components'
