--- app/src/atoms/buttons/OnDeviceDisplay/index.ts
@@ -1,6 +1,0 @@
-export { FloatingActionButton } from './FloatingActionButton'
-export { LargeButton } from './LargeButton'
-export { MediumButton } from './MediumButton'
-export { RadioButton } from './RadioButton'
-export { SmallButton } from './SmallButton'
-export { TabbedButton } from './TabbedButton'
