+++ app/src/atoms/buttons/SmallButton.tsx
@@ -12,7 +12,7 @@ import {
   ALIGN_CENTER,
   JUSTIFY_CENTER,
 } from '@opentrons/components'
-import { StyledText } from '../../text'
+import { StyledText } from '../text'
 import { ODD_FOCUS_VISIBLE } from './constants'
 import type { IconName, StyleProps } from '@opentrons/components'
