+++ app/src/atoms/buttons/index.ts
@@ -1,5 +1,11 @@
-export { TertiaryButton } from './TertiaryButton'
-export { QuaternaryButton } from './QuaternaryButton'
-export { SubmitPrimaryButton } from './SubmitPrimaryButton'
-export { ToggleButton } from './ToggleButton'
 export { BackButton } from './BackButton'
+export { FloatingActionButton } from './FloatingActionButton'
+export { LargeButton } from './LargeButton'
+export { MediumButton } from './MediumButton'
+export { QuaternaryButton } from './QuaternaryButton'
+export { RadioButton } from './RadioButton'
+export { SmallButton } from './SmallButton'
+export { SubmitPrimaryButton } from './SubmitPrimaryButton'
+export { TabbedButton } from './TabbedButton'
+export { TertiaryButton } from './TertiaryButton'
+export { ToggleButton } from './ToggleButton'
