+++ app/src/molecules/CardButton/index.tsx
@@ -14,7 +14,7 @@ import {
   JUSTIFY_CENTER,
 } from '@opentrons/components'
 import { StyledText } from '../../atoms/text'
-import { ODD_FOCUS_VISIBLE } from '../../atoms/buttons/OnDeviceDisplay/constants'
+import { ODD_FOCUS_VISIBLE } from '../../atoms/buttons/constants'
 
 import type { IconName } from '@opentrons/components'
