+++ app/src/organisms/InstrumentInfo/index.tsx
@@ -17,7 +17,7 @@ import {
 import { PipetteWizardFlows } from '../../organisms/PipetteWizardFlows'
 import { GripperWizardFlows } from '../../organisms/GripperWizardFlows'
 import { StyledText } from '../../atoms/text'
-import { MediumButton } from '../../atoms/buttons/OnDeviceDisplay'
+import { MediumButton } from '../../atoms/buttons'
 import { FLOWS } from '../PipetteWizardFlows/constants'
 import { formatTimestamp } from '../Devices/utils'
 import { GRIPPER_FLOW_TYPES } from '../GripperWizardFlows/constants'
