+++ app/src/organisms/LabwarePositionCheck/JogToWell.tsx
@@ -32,7 +32,7 @@ import { getIsOnDevice } from '../../redux/config'
 import { Portal } from '../../App/portal'
 import { ModalShell } from '../../molecules/Modal'
 import { StyledText } from '../../atoms/text'
-import { SmallButton } from '../../atoms/buttons/OnDeviceDisplay'
+import { SmallButton } from '../../atoms/buttons'
 import { NeedHelpLink } from '../CalibrationPanels'
 import { JogControls } from '../../molecules/JogControls'
 import { LiveOffsetValue } from './LiveOffsetValue'
