+++ app/src/organisms/OnDeviceDisplay/Navigation/index.tsx
@@ -18,7 +18,7 @@ import {
   DIRECTION_COLUMN,
 } from '@opentrons/components'
 
-import { ODD_FOCUS_VISIBLE } from '../../../atoms/buttons/OnDeviceDisplay/constants'
+import { ODD_FOCUS_VISIBLE } from '../../../atoms/buttons/constants'
 import { getLocalRobot } from '../../../redux/discovery'
 import { NavigationMenu } from './NavigationMenu'
