+++ app/src/organisms/PipetteWizardFlows/ExitModal.tsx
@@ -9,7 +9,7 @@ import {
   SecondaryButton,
   AlertPrimaryButton,
 } from '@opentrons/components'
-import { SmallButton } from '../../atoms/buttons/OnDeviceDisplay'
+import { SmallButton } from '../../atoms/buttons'
 import { SimpleWizardBody } from '../../molecules/SimpleWizardBody'
 import { FLOWS } from './constants'
 import type { PipetteWizardFlow } from './types'
