+++ app/src/pages/OnDeviceDisplay/Welcome.tsx
@@ -10,7 +10,7 @@ import {
   TYPOGRAPHY,
 } from '@opentrons/components'
 import { StyledText } from '../../atoms/text'
-import { SmallButton } from '../../atoms/buttons/OnDeviceDisplay'
+import { SmallButton } from '../../atoms/buttons'
 
 import screenImage from '../../assets/images/on-device-display/welcome_background.png'
